+++ src/components/UI/Footer.tsx
@@ -1,6 +1,6 @@
 const Footer = () => {
     return ( 
-      <div className = "footer secondary-txt"> 
+      <div className='footer secondary-txt'> 
       ©2022 Merhan Elkheshen 
       </div>
     );
