+++ src/components/UI/Header/Header.tsx
@@ -1,12 +1,12 @@
-import logo from '../../assets/logo.svg';
+import logo from '../../../assets/logo.svg';
 import { MdDarkMode, MdLightMode } from 'react-icons/md';
 import './Header.scss';
 
 const Header = ({theme, onToggleTheme}: any) => {
   return (
-    <div className="header">
-      <div className="logo-wrapper">
-        <img src={logo} alt="Logo" />
+    <div className='header'>
+      <div className='logo-wrapper'>
+        <img src={logo} alt='Logo' />
         <h2>Wallet Tracker</h2>
       </div>
         <span
